requirements.txt
beautifulsoup4==4.5.1
boto==2.39.0
defusedxml==0.4.1
Django==1.9.4
django-admin-tools==0.7.2
django-allauth==0.24.1
django-appconf==1.0.2
django-autoslug==1.9.3
django-bootstrap3==7.0.1
django-braces==1.8.1
django-crispy-forms==1.6.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
django-floppyforms==1.6.1
django-imagekit==3.3
django-mailgun==0.8.0
django-model-utils==2.4
django-redis==4.3.0
django-secure==1.0.1
django-storages-redux==1.3.2
easy-thumbnails==2.3
feedparser==5.2.1
gevent==1.1rc5
greenlet==0.4.9
gunicorn==19.4.5
html5lib==1.0b8
httplib2==0.9.2
oauthlib==1.0.3
pilkit==1.1.13
Pillow==3.1.1
pisa==3.0.33
psycopg2==2.6.1
PyPDF2==1.26.0
python3-openid==3.0.9
pytz==2015.7
redis==2.10.5
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
sqlparse==0.1.19
unicode-slugify==0.1.3
Unidecode==0.4.19
whitenoise==2.0.6
xhtml2pdf==0.1a4
